Move arm nodes to new /src folder.
