+++ launch/arm_teach_repeat_6_dof.launch
@@ -18,5 +18,5 @@
     home_position: [ 0,1.69517,2.16275,2.44159,-1.39,1.57459 ]
   </rosparam>
 
-  <node pkg="hebi_cpp_api_examples" type="teach_repeat_node" name="teach_repeat" output="screen"/>
+  <node pkg="hebi_cpp_api_examples" type="arm_teach_repeat_node" name="teach_repeat" output="screen"/>
 </launch>
