+++ src/kits/arm/arm_teach_repeat_node.hpp
